+++ .prettierrc
@@ -17,7 +17,6 @@
     "^@utils/",
     "^@context/",
     "^@lib/",
-    "^@types/",
     "^@styles/",
     "^@images/",
     "^@icons/*"
