+++ tsconfig.json
@@ -18,7 +18,6 @@
       "@context/*": ["./src/context/*"],
       "@lib/*": ["./src/lib/*"],
       "@styles/*": ["./src/styles/*"],
-      "@types/*": ["./src/lib/types/*"],
       "@images/*": ["./public/images/*"],
       "@icons/*": ["./public/icons/*"],
       "@utils/*": ["./src/utils/*"]
